Refactor css files and add mobile designs
