+++ src/components/JoinButton/style.scss
@@ -1,0 +1,3 @@
+.game-btn {
+    width: 140px;
+}
