+++ src/components/QuestionCard/style.scss
@@ -1,0 +1,6 @@
+.question-card {
+    &.ant-card .ant-card-head-title {
+        text-align: left;
+        font-size: x-large;
+    }
+}
