+++ src/components/SplashCard/style.scss
@@ -1,0 +1,6 @@
+.splash-card {
+    color: white;
+    font-size: xx-large;
+    font-family: "Nunito";
+    font-weight: bold;
+}
