+++ src/index.tsx
@@ -5,7 +5,6 @@ import App from "./App";
 
 import { BrowserRouter } from "react-router-dom";
 
-// eslint-disable-next-line no-undef
 const rootElement = document.getElementById("root")!;
 const root = createRoot(rootElement);
