+++ src/layouts/FlashCard.scss
@@ -1,0 +1,3 @@
+.ant-card {
+    margin: 5px;
+}
