+++ src/socket.ts
@@ -1,4 +1,3 @@
-/* eslint-disable no-undef */
 import { io } from "socket.io-client";
 
 // "undefined" means the URL will be computed from the `window.location` object
